ai-service/requirements.txt
fastapi
pydantic
pandas
numpy
scikit-learn
python-dotenv
google-genai
